A paint program that is able to save and load files.
This was a school assignment where we had to make a paint program given
by the professor, have a save and load file function. Although we had to
make this program before this project, because the professor wanted
everyone to have an equal footing with the program, it was necessary for
everyone to start with the same paint program. Nevertheless, the program
is able to save and load quite well, the saving function was done by
Shriram Anpalakan, and the loading function was done by me;.